Add tedge_core crate
This patch adds a first implementation of the tedge_core crate.

Its implementation uses an internal `Reactor` type for hiding the actual
implementation from the TedgeApplication type. This is not strictly
necessary, but gives a bit more seperation of concerns with
configuration vs "running" implementation.

The internal buffer size for the communication buffers between
core<->plugin handler is made configurable, so that a user can fine-tune
the core in this regard.

The core itself is implemented as a `CoreTask` type, which handles the
actual message-passing. All messages are routed over this CoreTask.

This ensures that we do only have 3 communication channels in each
PluginTask:

* One for CoreTask -> PluginTask
* One for PluginTask -> CoreTask
* One for Plugin -> PluginTask
(PluginTask -> Plugin is a function call: `Plugin::handle_message()` and
thus does not need a communication channel!)

and we thus do not run into the issue of having n^2 communication
channels for the communication between the PluginTask objects.

Tracing instrumentation is added to be able to have nice logging.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>