Add cancellation infrastructure in api
This patch adds cancellation code in the `Comms` type.

This interface can be used to receive "cancellation requests", that
might be issued by the core implementation in different situations (for
example on Ctrl-C).

For more information and what this entails, see the comment that was
added on the `CoreCommunication::cancellation_token()` function.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>