Add cancellation handling in application implementation
This patch makes use of the cancellation api of the
`tedge_api::CoreCommunication` type to implement cancellation of
plugins.

Note that the `PluginTask` does not know about cancellation because it
will automatically stop if the message receiver is closed, which might
or might not happen on a cancellation request.

If the cancellation request is issued, but plugins do not completely
halt their operation (just cancel what they are doing at the moment),
it is critical that the `CoreTask` (the routing part of the software)
does not stop working.

The cancellation token is passed to the plugins via the `Comms` type, so
the `PluginTask` does also not need to know about cancellation for the
same reasons.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>