Differentiate between plugin cancellation and task cancellation
This patch changes the cancellation setup in a way so that cancelling
the core (effectively shutting down) and cancelling a (didcated) plugin
is differentiated.

The `CancellationToken` that can be used to cancel a plugin is retreived
from a `CancellationToken` that is used for cancelling the `PluginTask`
(the task that orchestrates the `Plugin` inside the core).

When the whole application is cancelled, all these tokens fire. The
`PluginTask` has to make sure that the current message handling is
processed until it is ready and then return and shutdown the `Plugin`
instance cleanly.

There is no timeout for this added, because what would a timeout do?
Simply drop the futures (either the one that represents the last
`Plugin::handle_message()` call, or the one that represents the
`Plugin::shutdown()` call). Dropping these futures is the same as
killing the respective operation, so there's no benefit over just
waiting for the service-manager to kill the whole application.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>