Let CoreCommunication::send() return message id
This patch changes the `CoreCommunication::send()` signature, to return
the ID of the message.

This change makes the id of a message a public thing.

This is required so that when we wrap `CoreCommunication` to extend its
functionality, we can record the message id. As the `Message` object was
constructed inside `CoreCommunication::send()` before, this was not
possible.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>