Redefine common backend interface
This patch rewrites the backend interface that is common over all
backends to reflect the changes that were introduced with the typed API.

The diff here is that we do no longer have tmhe CoreCommunication type,
but pass list of addresses to the backend(s).
The list of addresses is put into an Arc so that we do not have to clone
lots of `Sender<_>` objects.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>
Signed-off-by: Marcel Müller <m.mueller@ifm.com>