Add implementation for timeout in plugin shutdown
This patch adds the implementation for timeouting plugin shutdown
routines.

This is required if a plugin behaves badly and `loop{}`s in its shutdown
implementation (either by programmer error or because it is waiting for
some resource to close for example).

By giving the user a configurable shutdown timeout, thin-edge can "kill"
plugins if their shutdown procedure takes too long.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>
Tested-by: Matthias Beyer <matthias.beyer@ifm.com>
Signed-off-by: Marcel Müller <m.mueller@ifm.com>