Add MQTT plugin and bridge plugin for it
This patch adds a MQTT plugin as well as a plugin that bridges
Measurement messages to the MQTT plugin.

The implementation is based on paho_mqtt.

The `OutgoingMessage` type is introduced, which gets handled by the
MQTT plugin.
This means that using the plugin will leak mqtt details into the
codebase of the using plugin.
As the using plugin is (or rather: should be) almost always a "bridge"
plugin, that bridges the actual message to an `OutgoingMessage` object
(such as the introduces Measurement bridge plugin), this is not
considered an issue.

For the reconnect logic, an exponential backoff is added via a crate
that implements such functionaly.
It starts with 100ms and increases up until 5s.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>