Make outgoing messages well-typed
This patch removes the genericness from the implementation of the mqtt
plugin, making the API of the plugin itself well-typed.

The `OutgoingMessage` type is introduced, which gets handled by the
plugin.

This means that using the plugin will leak mqtt details into the
codebase of the using plugin.
It is definitively not nice, but the implementation is _way_ less
complex to understand, and thus this is considered a benefit.

The configuration options for topic, qos and retain were removed, as
these are now encoded in the `OutgoingMessage` type.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>