Fix: Refresh sys stats before fetching them
We forgot that the sys stats that we read from the system needs to be
refresh()ed with this crate, before we get the current values.

This patches adds these calls to the best of my knowledge.
In the "load" backend, we `refresh_cpu()`, as I am assuming that this is
the right call, as there is no dedicated call like `refresh_load()`.
Also, as there is no `refresh_networks()` or similar, I am assuming that
the network interfaces of the sysinfo crate always use the latest
values.

Only the CPU backend changes were tested.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>
Tested-by: Matthias Beyer <matthias.beyer@ifm.com>