Add plugin_measurement_filter
This patch initially adds a "measurement_filter" plugin, which can be
used to extract values from measurement messages and apply a filter
function on them.

If the filter matches, the plugin forwards the message to a certain
other plugin, if not it can optionally send the message to an
alternative plugin.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>