Adapt core for PluginDirectory::get_address_for_self()
This patch implements the change for the
`PluginDirectory::get_address_for_self()` API.

The trick here is that we're using the previously implemented
`PluginDirectory` as a `CorePluginDirectory`, which implements all the
functionality except for the `get_address_for_self()` API.

A new type (conveniently named `PluginDirectory`) is introduced that
adds the implementation for the `get_address_for_self()` function on top
of the already standing implementation.
An instance of this type is only made for the call to the
`PluginBuilder::instantiate()` function.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>