Change config loading to support multiple files
This patch changes the configuration code to be able to load
configuration from multiple files.

That means that when a plugin is configured with

```toml
configuration = "some/file/path.toml"
```

for example, the configuration for that plugin is read from the
`some/file/path.toml` file.

The file path is not yet considered relative to the "root" configuration
file. This will be implemented in an additional patch.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>