Add ReplySender::closed()
This patch adds a function to check whether the ReplySender was closed.
This means that the plugin that sent the initial message stopped waiting
for a reply.

This can be useful for building a pipe-through like plugin, that
receives messages that it sends to another plugin and only executes a
side effect for each message.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>