Implement reply handling
This patch implements reply handling for the log plugin, making it able
to send replies from the plugin it pipes messages to back to the
original plugin.

This gives the user the possibility to add the log plugin between any
two plugins (as long as the messages they exchange implement `Debug`)
for debug-printing all messages exchanged.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>