Revert "Merge branch 'feature/add_tedge_api/log-pipe' into feature/add_tedge_api_impl"
We merged this in the believe that we implemented everything properly.
It turns out, that this was not the case.

What our problem here is, is that the `MessageBundle` that the
`LogPlugin` can receive does not necessarily implement `ReceiveBundle`,
which is what the `LogPlugin` would send out if it behaves like a pipe.

This would also require that there is a LogPlugin instanted for every
message you'd like to log in your configuration, which is clearly not
possible.

Hence, this whole idea was not working in the first place and should not
be implemented.

This reverts commit 8556a3e7d815bb1200d0a56b79d917ee38b78e07, reversing
changes made to 80a9221df136fdca22e39765f0cd114914cd1773.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>
Signed-off-by: Marcel Müller <m.mueller@ifm.com>