Implement relative config file path reading
This patch fixes the multi-config-file setup to be able to read
configuration files relative to the "root configuration file" (the one
that is passed via CLI).

The root config file path is passed to the verification routines which
then either reads the configuration of a plugin from that file, or, if
configured, finds the dedicated file and reads the configuration from
there.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>