Redesign tedge_core error handling strategy
We moved the API to miette in
commit 9ccd86589db8f9a72c4d7f545813e8da40b4d039 ("Replace PluginError with miette::Error").

In commit 909ed755e78eff256ae3efa75ce3a68d8fcabdd2 ("tedge_core: Adapt for miette as error handling crate")
we moved the tedge_core library to be able to handle the new API
interfaces based on miette.

With this patch, we continue the implementation of miette as error
handling library by redesigning the error handling story of tedge_core
to actually make use of the miette features. We do not wrap miette
errors from underlying plugins in a generic
TedgeApplicationError::Plugin that essentially loses all information
that miette carries, but actually use context mechanisms of miette to
annotate the error stack with further information.

Because of this, the public functions in this library were rewritten to
return `miette::Result` instead of `Result<_, TedgeApplicationError>`.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>