Impl concurrent message handling
This patch implements concurrent message handling.

In the previous implementation, we had the problem that all messages sent to a plugin were
not handled _concurrently_.
We simply looped over `self.plugin_msg_receiver.recv()`ed messages and passed them to
`self.plugin.handle_message()`, we did not get concurrency, but processed messages
sequentially.

For this to resolve, we use the following pattern: Two tasks are started:

* The first task receives messages in a loop. For each received
  message, a tokio task is spawned that calls `Plugin::handle_message()` (plus some panic
  catching). This task (its JoinHandle) is then pushed to a channel.
* The second task receives from the channel and puts these JoinHandle objects into a
  FuturesUnordered, which gets awaited in a streaming-like fashion (`StreamExt::next()`).
  This also happens to be in a loop, of course.

Because the changes are a bit involved here, we also seperated the
plugin lifecycle into three functions: one for plugin setup, one for
"main loop", and one for plugin shutdown.

Suggested-by: Marcel Müller <m.mueller@ifm.com>
Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>