Fix: tedge_core should return concrete error types
This patch fixes error handling in the tedge_core implementation once
again. Before this patch, we returned `miette::Result<_>` from the
interface of this library.

This should not have been the case. In fact, we can return concrete
error types in all cases, which is exactly what should be done.

This patch fixes this.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>
Suggested-by: Marcel Müller <m.mueller@ifm.com>