Redesign PluginTask::run() impl
This patch rewrites the PluginTask::run() implementation once again.

After the patches from
commit c0367fe348bae2dce36dc86838791edb0aa9eb39 ("Impl concurrent message handling")
where we introduced a pattern to make sure patches are handled
concurrently, we figured that waiting for the JoinHandle from
tokio::spawn() isn't actually necessary, because we're moving the Plugin
in a Arc<RwLock<_>> around, where the RwLock<_> automatically takes care
of the concurrency.

So instead, we can remove the whole multi-task waiting setup and rely on
the RwLock locking alone to make sure we properly wait for all messages
to be handled before shutting down the plugin.

Fixes: c0367fe348bae2dce36dc86838791edb0aa9eb39 ("Impl concurrent message handling")
Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>
Suggested-by: Marcel Müller <m.mueller@ifm.com>