Add Test for concurrently handling messages
The core is supposed to handle messages for plugins concurrently, this
test makes sure to verify this by sending a large amount of messages in
a short time. If 'awaited' concurrently, the handler futures will only
spend a total fraction of time of what a sequential handling would take.

A timeout of 400ms was chosen as it should be sufficient even on slower
machines.

Signed-off-by: Marcel Müller <m.mueller@ifm.com>