Add regression test: Fetching own address via name
Fetching the address of a plugin by name
(`PluginDirectory::get_address_for()`) should not result in an infinite
loop if that name is the name of the Plugin instance fetching the
address. In short: If a Plugin named "a" tries to fetch the address of a
Plugin named "a", we shouldn't loop infinitely.

This test checks whether that infinite loop exists by implementing a
`PluginBuilder` that fetches the address of the plugin itself by name
and instantiates the `Plugin` after that. The `Plugin` sends the
"StopCore" message to the thinedge core in its `Plugin::setup()`
implementation.

If the core is stopped normally via that mechanism, everything worked.
If the core gets stopped via the cancellation token (after 1 sec in the
test), the core was still working on the instantiation (meaning most
certainly that we hit the infinite loop), so we fail the test.

Suggested-by: Marcel Müller <m.mueller@ifm.com>
Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>