plugin_mqtt_measurement_bridge: Adapt to Plugin::handle_message() interface change
In commit abab64c856c0c299b13fcc1857a5b78f449e4a9e ("Add AnyMessage &
AnyMessages")

The `ReplySender` was renamed to `ReplySenderFor` and its interface was
changed so that we can specify the plugin_mqtt_measurement_bridge
message type of the original instead of the message type of the reply.

This commit adapts the plugin_mqtt_measurement_bridge implementation for
this change.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>