Fix: MessageType is no longer Eq, Hash, use lists for support-check
This patch fixes the MessageType handling. In
commit abab64c856c0c299b13fcc1857a5b78f449e4a9e ("Add AnyMessage & AnyMessages")
we changed the message type handling so that we can use the new
`MessageType` type instead of raw `str` and `TypeId`.
Because this type cannot implement `Eq`, `PartialEq` and `Hash`, we can
no longer store this information in `HashSet`.

Thus, change the implementation to store this information in `Vec`.
Because of the small problem size we're talking about here (lower 10s),
this is okay.

Suggested-by: Marcel Müller <m.mueller@ifm.com>
Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>