Redesign logging setup
This patch rewrites the logging setup in the CLI implementation in a way
so that by default we get WARN and ERROR logging (which can be turned
off completely if desired).

If the user wants, they can specify `--logging info` for a bit more
verbosity, or `--logging debug` for a lot more verbosity.
If they feel confident, `--logging trace` drowns them in log output ;-)

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>