Add helper type for Humantime configuration members
This patch adds a helper type for having a human-readable duration in
the configuration of a plugin.
This helper type implements Deserialize with the humantime crate, so
that durations can be specified as (e.g.) "5mins".

The AsConfig trait is implemented so that it can be used in
self-describing configuration types.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>