Shrink response types
This patch shrinks the response types down to enums.

This is necessary because right now we cannot implement more than 10
handlers for a plugin, but implementing a decent showcase would require
more than that.

And because we don't want to touch the core implementation just for the
showcase, we simply shrink the number of types.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>