plugin_sysstat: Change config type to use typed address configuration
This patch changes the "sysstat" plugin to use the new
tedge_lib::address::AddressGroup interface for sending messages to
multiple targets at once.

This change is a bit more involved and also shows where the
`SendAllResult<_>` interface lacks: The `SendAllResult<_>::into()`
function must by type-annotated, and this clearly is not nice.

But this is a first step into the right direction.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>