Update core to use the new tedge_api interface
The tedge_api interface now no longer uses an MPSC channel but instead
requires the core to provide a 'MessageProvider' function. This function
is a `Fn` closure, meaning that it:

- Owns its contents, no borrows of any kinds (: 'static)
- Can be called many times

This closure has to return a BoxedFuture which when polled will spawn a
task that calls `handle_message` on the plugin it was built with.

Signed-off-by: Marcel Müller <m.mueller@ifm.com>