Add default bodies to tedge_api::Plugin
This patch adds default bodies for the `Plugin` trait methods, as some
plugins don't want to do anything in these functions, so they would
implement them with simply returning `Ok(())` anyways, which we can do
here equally good.

This results in less brain-load on the Plugin authors side as well.

Signed-off-by: Marcel Müller <m.mueller@ifm.com>
Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>
Reviewed-by: Matthias Beyer <matthias.beyer@ifm.com>