Add Plugin::main() with a default body
The plugin "main" should implement the main usecase of the plugin. This
is where plugin authors should do their work.

Of course, having a default implementation makes implementing a Plugin
that only handles messages as easy as it can get: An author just has to
implement an seemingly empty `Plugin` trait for their type, plus the
`Handle<>` implementations for message handling.

If they want to do extra things, overwriting the functions from `Plugin`
is the way to go.

Signed-off-by: Marcel Müller <m.mueller@ifm.com>
Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>
Reviewed-by: Matthias Beyer <matthias.beyer@ifm.com>