Do not necessarily pass &str anymore to fs helper fns
This patch changs the calls of the fs helper functions where the
interface of the functions was changed from taking `&str` as path
argument to `impl AsRef<Path>`.

Because of this, it is not necessary anymore that `&str` is passed, but
`Path`es can be passed or, as in most cases, the ad-hoc created
`String` objects can be passed by value instead of by reference,
resulting in less mental load on the calling side.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>