Change TypeUuid implementation to bevy_reflect
This patch replaces the type_uuid crate with the bevy_reflect
implementation of the functionality.

The rationale for this is, that the implementation of the type_uuid
crate does not support generics, but the bevy_reflect implementation
does.

As we need generic support for type uuids, we change to bevy_reflect
here.

Suggested-by: Marcel Müller <m.mueller@ifm.com>
Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>