Fix: Change Plugin impl to use main() fn
In commit
d6517ace8e343145c868d5aebbaa328ad0f880cb ("Add Plugin::main() with a default body")
we changed the Plugin trait to have a `main` function that should be
used to implement the main usecase of a plugin.

We did not change it here in the test, and because communication is not
enabled in the moment the `Plugin::start()` implementation is called,
this test failed.

This patch fixes this by moving the `Plugin::start()` implementation to
the `Plugin::main()` one.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>
Tested-by: Matthias Beyer <matthias.beyer@ifm.com>