Rename: communication_buffer_size -> max_concurrency
The "communication_buffer_size" name came from the old setup of the
communcation, where we used channels for the inter-plugin communication.
Because we changed this to be actually only async function calls with
some decoupling for panic handling.

"max_concurrency" is a better name, as we're only resticting the
concurrency here, not some size of some buffer or channel.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>