Revert "Let PluginDirectory::get_address_for() take AsRef<str>"
This patch actually broke tests and made the interface _harder_ to use,
because now we have to specify two generic arguments.

We revert this patch therefore. What we tried is using an
`impl AsRef<str>`, but because impl Trait cannot be used with generics
(see `rustc --explain E0632`), we revert the implementation to the
previous situation where we simply take an `&str` as name for the
address.

This reverts commit 24e720d8273733a612a283d53088be221d98bb00.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>