Fix: Use right configuration key
And this, my dear people, is why we need a merge bot.

In a1b96c0f7d39b214766079d6be80700c2f637c9d ("Rename: communication_buffer_size -> max_concurrency")
we renamed the configuration key. This was merged, but the branchoff for
the branch implementing this test was _before_ that mentioned merge.
It used the old configuration key, everything was fine... until the
test was merged, which resulted in the old key being used. The target
branch, though, already implemented the new key.

So it broke the tests.

This wouldn't have happened if a merge bot was in place here, which
would have checked whether the merged branch would pass CI and so on.

Maybe we get one, one day.

Signed-off-by: Matthias Beyer <matthias.beyer@ifm.com>